Prevent initial useEffect call in Feed component
Modified the Feed component to prevent the useEffect hook from running its logic on the initial component mount. Introduced a useRef hook to track the component's mount status. This change ensures that the getPosts function is only called in response to changes in postFeedSort or postFeedType, avoiding unnecessary API calls during the initial render.

https://github.com/sublinks/sublinks-frontend/issues/48